Move folding range
