+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/DefaultRazorLanguageServerCustomMessageTarget.cs
@@ -164,115 +164,6 @@ internal partial class DefaultRazorLanguageServerCustomMessageTarget : RazorLang
         return request?.Response;
     }
 
-    public override async Task<RazorFoldingRangeResponse?> ProvideFoldingRangesAsync(RazorFoldingRangeRequestParam foldingRangeParams, CancellationToken cancellationToken)
-    {
-        if (foldingRangeParams is null)
-        {
-            throw new ArgumentNullException(nameof(foldingRangeParams));
-        }
-
-        var csharpRanges = new List<FoldingRange>();
-        var csharpTask = Task.Run(async () =>
-        {
-            var (synchronized, csharpSnapshot) = await _documentSynchronizer.TrySynchronizeVirtualDocumentAsync<CSharpVirtualDocumentSnapshot>(
-                _documentManager, foldingRangeParams.HostDocumentVersion, foldingRangeParams.TextDocument, cancellationToken);
-
-            if (synchronized)
-            {
-                var csharpRequestParams = new FoldingRangeParams()
-                {
-                    TextDocument = foldingRangeParams.TextDocument.WithUri(csharpSnapshot.Uri),
-                };
-
-                var request = await _requestInvoker.ReinvokeRequestOnServerAsync<FoldingRangeParams, IEnumerable<FoldingRange>?>(
-                    csharpSnapshot.Snapshot.TextBuffer,
-                    Methods.TextDocumentFoldingRange.Name,
-                    RazorLSPConstants.RazorCSharpLanguageServerName,
-                    SupportsFoldingRange,
-                    csharpRequestParams,
-                    cancellationToken).ConfigureAwait(false);
-
-                var result = request?.Response;
-                if (result is null)
-                {
-                    csharpRanges = null;
-                }
-                else
-                {
-                    csharpRanges.AddRange(result);
-                }
-            }
-        }, cancellationToken);
-
-        var htmlRanges = new List<FoldingRange>();
-        var htmlTask = Task.CompletedTask;
-        htmlTask = Task.Run(async () =>
-        {
-            var (synchronized, htmlDocument) = await _documentSynchronizer.TrySynchronizeVirtualDocumentAsync<HtmlVirtualDocumentSnapshot>(
-                _documentManager, foldingRangeParams.HostDocumentVersion, foldingRangeParams.TextDocument, cancellationToken);
-
-            if (synchronized)
-            {
-                var htmlRequestParams = new FoldingRangeParams()
-                {
-                    TextDocument = new()
-                    {
-                        Uri = htmlDocument.Uri
-                    }
-                };
-
-                var request = await _requestInvoker.ReinvokeRequestOnServerAsync<FoldingRangeParams, IEnumerable<FoldingRange>?>(
-                    htmlDocument.Snapshot.TextBuffer,
-                    Methods.TextDocumentFoldingRange.Name,
-                    RazorLSPConstants.HtmlLanguageServerName,
-                    SupportsFoldingRange,
-                    htmlRequestParams,
-                    cancellationToken).ConfigureAwait(false);
-
-                var result = request?.Response;
-                if (result is null)
-                {
-                    htmlRanges = null;
-                }
-                else
-                {
-                    htmlRanges.AddRange(result);
-                }
-            }
-        }, cancellationToken);
-
-        try
-        {
-            await Task.WhenAll(htmlTask, csharpTask).ConfigureAwait(false);
-        }
-        catch (Exception)
-        {
-            // Return null if any of the tasks getting folding ranges
-            // results in an error
-            return null;
-        }
-
-        // Since VS FoldingRanges doesn't poll once it has a non-null result returning a partial result can lock us
-        // into an incomplete view until we edit the document. Better to wait for the other server to be ready.
-        if (htmlRanges is null || csharpRanges is null)
-        {
-            return null;
-        }
-
-        return new(htmlRanges.ToImmutableArray(), csharpRanges.ToImmutableArray());
-    }
-
-    private static bool SupportsFoldingRange(JToken token)
-    {
-        var serverCapabilities = token.ToObject<ServerCapabilities>();
-
-        var supportsFoldingRange = serverCapabilities?.FoldingRangeProvider?.Match(
-            boolValue => boolValue,
-            options => options is not null) ?? false;
-
-        return supportsFoldingRange;
-    }
-
     public override Task<WorkspaceEdit?> ProvideTextPresentationAsync(RazorTextPresentationParams presentationParams, CancellationToken cancellationToken)
     {
         return ProvidePresentationAsync(presentationParams, presentationParams.HostDocumentVersion, presentationParams.Kind, VSInternalMethods.TextDocumentTextPresentationName, cancellationToken);
